feat: add common ShareCard component
