+++ src/screens/InviteFlow/InviteShare/index.tsx
@@ -6,7 +6,7 @@ import {COLORS} from '@constants/colors';
 import {commonStyles, windowWidth} from '@constants/styles';
 import {Header} from '@navigation/components/Header';
 import {useFocusStatusBar} from '@navigation/hooks/useFocusStatusBar';
-import {ShareCard} from '@screens/InviteFlow/InviteShare/components/ShareCard';
+import {InviteShareCard} from '@screens/InviteFlow/InviteShare/components/InviteShareCard';
 import {replaceString, t, tagRegex} from '@translations/i18n';
 import {font} from '@utils/styles';
 import React, {memo} from 'react';
@@ -37,7 +37,7 @@ export const InviteShare = memo(() => {
           )}
         </Text>
         <Image source={icon} style={styles.icon} />
-        <ShareCard />
+        <InviteShareCard />
       </View>
     </View>
   );
