+++ src/screens/InviteFlow/QRCodeShare/components/QRCodePreview/components/QRCode/components/QRCodeLogo.tsx
