+++ src/screens/InviteFlow/QRCodeShare/components/QRCodePreview/components/QRCode/constants.ts
